# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AuditOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List audit findings
      parameters:
        - name: engagementId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
    post:
      operationId: createFinding
      tags: [findings]
      summary: Create a new finding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFindingBody"
      responses:
        "201":
          description: Created finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
  /findings/{id}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get a finding by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update a finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFindingBody"
      responses:
        "200":
          description: Updated finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /working-papers:
    get:
      operationId: listWorkingPapers
      tags: [working-papers]
      summary: List working papers
      parameters:
        - name: engagementId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of working papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingPaper"
    post:
      operationId: createWorkingPaper
      tags: [working-papers]
      summary: Create a new working paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkingPaperBody"
      responses:
        "201":
          description: Created working paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkingPaper"
  /working-papers/{id}:
    get:
      operationId: getWorkingPaper
      tags: [working-papers]
      summary: Get a working paper by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Working paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkingPaper"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWorkingPaper
      tags: [working-papers]
      summary: Update a working paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkingPaperBody"
      responses:
        "200":
          description: Updated working paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkingPaper"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /working-papers/{id}/draft:
    post:
      operationId: draftWorkingPaper
      tags: [ai]
      summary: AI-draft content for a working paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI drafted content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Finding:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        anomalyId:
          type: ["integer", "null"]
        findingRef:
          type: string
        title:
          type: string
        description:
          type: string
        riskLevel:
          type: string
        isaReference:
          type: ["string", "null"]
        nbaaReference:
          type: ["string", "null"]
        managementResponse:
          type: ["string", "null"]
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        raisedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - findingRef
        - title
        - description
        - riskLevel
        - status
        - createdAt
        - updatedAt
    CreateFindingBody:
      type: object
      properties:
        engagementId:
          type: integer
        anomalyId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: string
        riskLevel:
          type: string
        isaReference:
          type: ["string", "null"]
        nbaaReference:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        raisedBy:
          type: ["integer", "null"]
      required:
        - engagementId
        - title
        - description
        - riskLevel
    UpdateFindingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        riskLevel:
          type: string
        managementResponse:
          type: ["string", "null"]
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
    WorkingPaper:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        wpRef:
          type: string
        section:
          type: string
        title:
          type: string
        aiDrafted:
          type: boolean
        contentText:
          type: ["string", "null"]
        status:
          type: string
        preparedBy:
          type: ["integer", "null"]
        preparedAt:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        reviewedAt:
          type: ["string", "null"]
        signedOffBy:
          type: ["integer", "null"]
        signedOffAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - wpRef
        - section
        - title
        - aiDrafted
        - status
        - createdAt
        - updatedAt
    CreateWorkingPaperBody:
      type: object
      properties:
        engagementId:
          type: integer
        wpRef:
          type: string
        section:
          type: string
        title:
          type: string
        contentText:
          type: ["string", "null"]
        preparedBy:
          type: ["integer", "null"]
      required:
        - engagementId
        - wpRef
        - section
        - title
    UpdateWorkingPaperBody:
      type: object
      properties:
        title:
          type: string
        contentText:
          type: ["string", "null"]
        status:
          type: string
        preparedBy:
          type: ["integer", "null"]
        preparedAt:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        reviewedAt:
          type: ["string", "null"]
        signedOffBy:
          type: ["integer", "null"]
        signedOffAt:
          type: ["string", "null"]
    AiDraftResult:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...